Langchain_models/Embedding_models/embed_gpt_query.py
# openai embedding models require paid api credits to this code is just for reference

from langchain_openai import OpenAIEmbeddings
from dotenv import load_dotenv

load_dotenv()

embedding = OpenAIEmbeddings(model='text-embedding-3-large', dimensions=32)
result = embedding.embed_query("Hello world")
print(str(result))