pyproject.toml
[project]
name = "langchain-models"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.6",
    "huggingface-hub>=0.36.0",
    "langchain>=1.2.0",
    "langchain-anthropic>=1.3.1",
    "langchain-core>=1.2.6",
    "langchain-google-genai>=4.1.3",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.6",
    "numpy>=2.4.0",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "transformers>=4.57.3",
]
